fixed the segfault error
